# openapi.yaml
openapi: 3.0.3
info:
  title: Pulse Authentication API
  version: 1.0.0
  description: >
    Authentication backend for Pulse. Covers signup, email verification (+ resend),
    login with IP/account lockout, logout, password reset (+ resend), change password,
    token rotation, and the authenticated profile endpoint.

    **Auth model:** on successful login / email verification / token rotation, the server
    sets two **httpOnly** cookies — `accessToken` and `refreshToken`. Protected endpoints
    read `accessToken` (and `rotateToken` reads `refreshToken`). All browser calls must send
    cookies (`credentials: "include"`).

    **Response envelope:** success responses use `{ statuscode, data, message, sucess }`
    (note: the success flag is spelled `sucess`). Error responses use
    `{ success: false, message, errors }`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /pulse/images/{id}:
    get:
      tags: [Images]
      summary: Fetch an uploaded image
      description: >
        Returns the raw image bytes by id. **Authenticated and owner-scoped** — the image
        must belong to one of the caller's cookie policies (joined via
        cookie_policy → websites → userId). A non-existent id OR another user's image both
        return `404` (so ids aren't enumerable); an unreadable S3 object also `404`s. The
        server reads the bytes from S3 and streams them directly (no presigned URL / redirect).
        Rendered in the editor/preview via the auto-attached `accessToken` cookie; served with
        `Cache-Control: private, max-age=31536000, immutable` (a given id maps to a fixed object).
      security:
        - accessTokenCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: The image id (uuid).
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The image binary.
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Image not found (or not owned by the caller).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                { success: false, message: 'image not found', errors: [] }
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: accessToken
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope from the global error handler.
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        errors:
          description: Empty array for most errors; an array of field errors for 422.
          example: []
      required: [success, message]
...